Retrieve data for a list of species from the ppendemic_tab dataset.

This function takes a list of species names, searches for their data in the ppendemic_tab dataset, and returns a data frame containing the relevant information for each species.

Usage

get_ppendemic_data(splist, max_distance = 0.1)

Arguments

splist

A character vector containing the names of the species to search for.

max_distance

The maximum allowed distance for fuzzy matching of species names. Defaults to 0.1.

Value

A data frame containing the retrieved information for each species.

Examples

get_ppendemic_data(splist = c("Aa aurantiaca", "Aa aurantiaaia", "werneria nubigena"))
#>      name_submitted accepted_name accepted_family accepted_name_author
#> 1     Aa aurantiaca Aa aurantiaca     Orchidaceae           D.Trujillo
#> 2    Aa aurantiaaia Aa aurantiaca     Orchidaceae           D.Trujillo
#> 3 Werneria nubigena          nill            nill                 nill
#>   publication_author place_of_publication volume_and_page first_published dist
#> 1               <NA>        Lankesteriana           11: 3          (2011)    0
#> 2               <NA>        Lankesteriana           11: 3          (2011)    2
#> 3               nill                 nill            nill            nill nill