pep_regecol

List all endemic species within an ecological region.

List of ecological region ids and full name, based on Zamora, C. 1996.

AA = Altoandina

BHA = Bosques Húmedos Amazónicos

BHMP = Bosques Muy Húmedos Premontanos

BMHM = Bosques Muy Húmedos Montanos

BMHP = Bosques Muy Húmedos Premontanos

BPM = Bosques Pluviales Montanos

BS = Bosques Secos

DCT = Desierto Cálido Tropical

DST = Desierto Semicálido Tropical

MA = Mesoandina

MDE = Matorral Desértico

NDE = Matorral Desértico

PAR = Páramo

PD = Puna Desértica

PSH = Puna Húmeda y Seca

pep_regecol(region = NULL, specie)

Arguments

region
Atomic element or vector with ecological region names, shrot or long name.

specie
Atomic element or vector with species names.

Value

A tibble

References

León, B., J. Roque, C. Ulloa Ulloa, N. C. A. Pitman, P. M. Jørgensen & A. Cano Echevarría. 2006 (2007). El Libro Rojo de las Plantas Endémicas del Perú. Revista Peruana Biol. 13(núm. 2 especial): 1s–971s

ZAMORA, C. 1996. Las regiones ecológicas del Perú. En: Rodríguez, L.O. (ed.), Diversidad Biológica del Perú: Zonas Prioritarias para su Conservación, pp. 137-141. FANPE, GTZ-INRENA. Lima, Perú.

Examples

# By Region ID
pep_regecol("BHA")

#> Region: BHA with 754 species

#> # A tibble: 756 x 3
#>    region_id region_eco                 accepted_name              
#>    <chr>     <chr>                      <chr>                      
#>  1 BHA       Bosques Húmedos Amazónicos Aphelandra cuscoensis      
#>  2 BHA       Bosques Húmedos Amazónicos Aphelandra latibracteata   
#>  3 BHA       Bosques Húmedos Amazónicos Justicia manserichensis    
#>  4 BHA       Bosques Húmedos Amazónicos Mendoncia peruviana        
#>  5 BHA       Bosques Húmedos Amazónicos Mendoncia tarapotana       
#>  6 BHA       Bosques Húmedos Amazónicos Odontophyllum cuscoensis   
#>  7 BHA       Bosques Húmedos Amazónicos Oplonia hutchisonii        
#>  8 BHA       Bosques Húmedos Amazónicos Pachystachys killipii      
#>  9 BHA       Bosques Húmedos Amazónicos Pachystachys incarnata     
#> 10 BHA       Bosques Húmedos Amazónicos Pachystachys longibracteata
#> # … with 746 more rows

pep_regecol(c("MA", "BS"))

#> Regions: BS - MA with 1451 species

#> # A tibble: 1,468 x 3
#>    region_id region_eco    accepted_name             
#>    <chr>     <chr>         <chr>                     
#>  1 MA        Mesoandina    Aphelandra juninensis     
#>  2 MA        Mesoandina    Aphelandra viscosa        
#>  3 MA        Mesoandina    Dicliptera porphyrea      
#>  4 MA        Mesoandina    Justicia weberbaueri      
#>  5 BS        Bosques Secos Justicia weberbaueri      
#>  6 MA        Mesoandina    Justicia alpina           
#>  7 BS        Bosques Secos Tetramerium denudatum     
#>  8 BS        Bosques Secos Tetramerium peruvianum    
#>  9 BS        Bosques Secos Tetramerium sagasteguianum
#> 10 BS        Bosques Secos Tetramerium surcubambense 
#> # … with 1,458 more rows

# By species name
pep_regecol(specie = "Aphelandra weberbaueri")

#> # A tibble: 2 x 3
#>   region_id region_eco                      accepted_name         
#>   <chr>     <chr>                           <chr>                 
#> 1 BMHM      Bosques Muy Húmedos Montanos    Aphelandra weberbaueri
#> 2 BMHP      Bosques Muy Húmedos Premontanos Aphelandra weberbaueri

especie <-  c("Aphelandra weberbaueri", "Odontophyllum cuscoensis")
pep_regecol(specie = especie)

#> # A tibble: 3 x 3
#>   region_id region_eco                      accepted_name           
#>   <chr>     <chr>                           <chr>                   
#> 1 BMHM      Bosques Muy Húmedos Montanos    Aphelandra weberbaueri  
#> 2 BMHP      Bosques Muy Húmedos Premontanos Aphelandra weberbaueri  
#> 3 BHA       Bosques Húmedos Amazónicos      Odontophyllum cuscoensis